+++ src/plugins/sensor/Sensor_AmbientLight.cpp
@@ -50,6 +50,7 @@ Sensor_AmbientLight::~Sensor_AmbientLight()
 MQTT::AutoDiscovery::EntityPtr Sensor_AmbientLight::getAutoDiscovery(FormatType format, uint8_t num)
 {
     auto discovery = new MQTT::AutoDiscovery::Entity();
+    __DBG_discovery_printf("num=%u/%u d=%p", num, getAutoDiscoveryCount(), discovery);
     auto baseTopic = MQTT::Client::getBaseTopicPrefix();
     switch(num) {
         case 0: {