+++ src/plugins/sensor/Sensor_DHTxx.cpp
@@ -28,6 +28,7 @@ Sensor_DHTxx::~Sensor_DHTxx()
 MQTT::AutoDiscovery::EntityPtr getAutoDiscovery(FormatType format, uint8_t num)
 {
     auto discovery = new MQTT::AutoDiscovery::Entity();
+    __DBG_discovery_printf("num=%u/%u d=%p", num, getAutoDiscoveryCount(), discovery);
     switch(num) {
         case 0:
             if (discovery->create(this, _getId(FSPGM(temperature)), format)) {